requirements.txt

async-generator
attrs
beautifulsoup4
certifi
cffi
distlib
dj-database-url
Django
django-bootstrap-form
django-bootstrap4
django-on-heroku
filelock
gunicorn
h11
idna
mod-wsgi
MouseInfo
outcome
platformdirs
psycopg
psycopg2
psycopg2-binary
PyAutoGUI
pycparser
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
PySocks
python-dotenv
pytweening
selenium
sniffio
sortedcontainers
soupsieve
sqlparse
trio
trio-websocket
tzdata
urllib3
virtualenv
whitenoise
wsproto
django-recaptcha
